Yay testing framework!
So yesterday I added a commit to make sure no NAs sneak into generated
count tables due to inappropriately read tables with/without headers.
What I did not realize is that if there are no NAs, then that reasonably
smart piece of logic will fail because it asks for an index of a null
value, evaluating to logical(0).  I then test this value against the
rownames column in the data, but since logical(0) is of a different
type, it in turn evaluates to a null index of rows to keep.  I of course
then subset the data on that index, resulting in 0 kept rows.

So long story short, tests are good because they check for stuff that I
will never think to check for.