playing with isva/smartsva/ica/tsne/other pca methods
I did a little fooling around today at the machine learning workshop.
In that process, I started finally cleaning up my PCA plotter; but now
it should probably be renamed to something like
reduce_dimensionality_plot or something odd, as it at least in theory is
now able to use a bunch of pca methods, ica, and tsne.

In a similar fashion, I was playing with isva/smartsva in the hopes of
finding ways to query/poke at confounded data.