Significant documentation changes, with a problem
I documented all the stuff in 'inst/' in the hope of making it easier to
use the examples in the docstrings as well as in the hopes of making
this package a bit more accessible.  Unfortunately, my various
invocations of system.file(package="hpgltools") in 01_hpgltools.r
cause the installation to fail now, unless one uses
'--no-staged-install'.  For the moment, I will leave this as-is, but
will check in with Theresa if it works for her.